gemini-backend/backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
python-multipart
PyPDF2
python-docx